requirements.txt
Django
psycopg2
djangorestframework
django-machina
Pillow
